+++ src/flatten/tests.js
@@ -20,3 +20,10 @@ test('Test array of complex objects', t => {
     thing: 1
   }]);
 });
+
+test('Test Multi level of arrays', t => {
+  const results = flatten([1, 2, [3, 4], 5, [6, [7, 8, [9, [10, 11], 12]]]]);
+
+  t.truthy(results);
+  t.deepEqual(results, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
+});