+++ src/last/tests.js
@@ -1,11 +1,13 @@
 import last from './index';
 import test from 'ava';
 
-test('test last() functionality', t => {
+test('test last() Array functionality', t => {
   const list = [1, 2, 3, 4];
 
   t.is(last(list), 4);
+});
 
+test('test last() String functionality', t => {
   const str = 'abcd';
 
   t.is(last(str), 'd');