+++ src/omit/index.js
@@ -1,5 +1,11 @@
 import isArray from '../isArray';
 
+// Check if the key is an array and find the current prop within it if so
+const findKey = (key, prop) => isArray(key) && key.indexOf(prop) === -1;
+
+// Check if the key is NOT an array and make sure it isn't equal to our current prop
+const compareKey = (key, prop) => !isArray(key) && key !== prop;
+
 /**
  * @description Create a new Array/Object by omitting the requested values
  * @param  {String} key The key(s) in which to omit from the data
@@ -12,8 +18,10 @@ import isArray from '../isArray';
  */
 const omit = (key, x) => {
   if (isArray(x)) {
+    const keysAreArr = isArray(key);
+
     return x.filter(val => {
-      if (isArray(key)) {
+      if (keysAreArr) {
         return key.indexOf(val) === -1;
       }
 
@@ -24,13 +32,13 @@ const omit = (key, x) => {
   const keys = Object.keys(x);
 
   return keys.reduce((acc, prop) => {
-    if (isArray(key) && key.indexOf(prop) === -1) {
+    if (findKey(key, prop)) {
       acc[prop] = x[prop];
 
       return acc;
     }
 
-    if (!isArray(key) && key !== prop) {
+    if (compareKey(key, prop)) {
       acc[prop] = x[prop];
     }